+++ src/test/java/wooteco/subway/AcceptanceTest.java
@@ -26,347 +26,347 @@ import wooteco.subway.service.station.dto.StationResponse;
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
 @Sql("/truncate.sql")
 public class AcceptanceTest {
-    public static final String STATION_NAME_KANGNAM = "강남역";
-    public static final String STATION_NAME_YEOKSAM = "역삼역";
-    public static final String STATION_NAME_SEOLLEUNG = "선릉역";
-    public static final String STATION_NAME_HANTI = "한티역";
-    public static final String STATION_NAME_DOGOK = "도곡역";
-    public static final String STATION_NAME_MAEBONG = "매봉역";
-    public static final String STATION_NAME_YANGJAE = "양재역";
+	public static final String STATION_NAME_KANGNAM = "강남역";
+	public static final String STATION_NAME_YEOKSAM = "역삼역";
+	public static final String STATION_NAME_SEOLLEUNG = "선릉역";
+	public static final String STATION_NAME_HANTI = "한티역";
+	public static final String STATION_NAME_DOGOK = "도곡역";
+	public static final String STATION_NAME_MAEBONG = "매봉역";
+	public static final String STATION_NAME_YANGJAE = "양재역";
 
-    public static final String LINE_NAME_2 = "2호선";
-    public static final String LINE_NAME_3 = "3호선";
-    public static final String LINE_NAME_BUNDANG = "분당선";
-    public static final String LINE_NAME_SINBUNDANG = "신분당선";
+	public static final String LINE_NAME_2 = "2호선";
+	public static final String LINE_NAME_3 = "3호선";
+	public static final String LINE_NAME_BUNDANG = "분당선";
+	public static final String LINE_NAME_SINBUNDANG = "신분당선";
 
-    public static final String TEST_USER_EMAIL = "brown@email.com";
-    public static final String TEST_USER_NAME = "브라운";
-    public static final String TEST_USER_PASSWORD = "brown";
+	public static final String TEST_USER_EMAIL = "brown@email.com";
+	public static final String TEST_USER_NAME = "브라운";
+	public static final String TEST_USER_PASSWORD = "brown";
 
-    @LocalServerPort
-    public int port;
+	@LocalServerPort
+	public int port;
 
-    public static RequestSpecification given() {
-        return RestAssured.given().log().all();
-    }
+	public static RequestSpecification given() {
+		return RestAssured.given().log().all();
+	}
 
-    @BeforeEach
-    public void setUp() {
-        RestAssured.port = port;
-    }
+	@BeforeEach
+	public void setUp() {
+		RestAssured.port = port;
+	}
 
-    public StationResponse createStation(String name) {
-        Map<String, String> params = new HashMap<>();
-        params.put("name", name);
+	public StationResponse createStation(String name) {
+		Map<String, String> params = new HashMap<>();
+		params.put("name", name);
 
-        return
-            given().
-                body(params).
-                contentType(MediaType.APPLICATION_JSON_VALUE).
-                accept(MediaType.APPLICATION_JSON_VALUE).
-                when().
-                post("/stations").
-                then().
-                log().all().
-                statusCode(HttpStatus.CREATED.value()).
-                extract().as(StationResponse.class);
-    }
+		return
+			given().
+				body(params).
+				contentType(MediaType.APPLICATION_JSON_VALUE).
+				accept(MediaType.APPLICATION_JSON_VALUE).
+				when().
+				post("/stations").
+				then().
+				log().all().
+				statusCode(HttpStatus.CREATED.value()).
+				extract().as(StationResponse.class);
+	}
 
-    public List<StationResponse> getStations() {
-        return
-            given().when().
-                get("/stations").
-                then().
-                log().all().
-                extract().
-                jsonPath().getList(".", StationResponse.class);
-    }
+	public List<StationResponse> getStations() {
+		return
+			given().when().
+				get("/stations").
+				then().
+				log().all().
+				extract().
+				jsonPath().getList(".", StationResponse.class);
+	}
 
-    public void deleteStation(Long id) {
-        given().when().
-            delete("/stations/" + id).
-            then().
-            log().all();
-    }
+	public void deleteStation(Long id) {
+		given().when().
+			delete("/stations/" + id).
+			then().
+			log().all();
+	}
 
-    public LineResponse createLine(String name) {
-        Map<String, String> params = new HashMap<>();
-        params.put("name", name);
-        params.put("startTime", LocalTime.of(5, 30).format(DateTimeFormatter.ISO_LOCAL_TIME));
-        params.put("endTime", LocalTime.of(23, 30).format(DateTimeFormatter.ISO_LOCAL_TIME));
-        params.put("intervalTime", "10");
+	public LineResponse createLine(String name) {
+		Map<String, String> params = new HashMap<>();
+		params.put("name", name);
+		params.put("startTime", LocalTime.of(5, 30).format(DateTimeFormatter.ISO_LOCAL_TIME));
+		params.put("endTime", LocalTime.of(23, 30).format(DateTimeFormatter.ISO_LOCAL_TIME));
+		params.put("intervalTime", "10");
 
-        return
-            given().
-                body(params).
-                contentType(MediaType.APPLICATION_JSON_VALUE).
-                accept(MediaType.APPLICATION_JSON_VALUE).
-                when().
-                post("/lines").
-                then().
-                log().all().
-                statusCode(HttpStatus.CREATED.value()).
-                extract().as(LineResponse.class);
-    }
+		return
+			given().
+				body(params).
+				contentType(MediaType.APPLICATION_JSON_VALUE).
+				accept(MediaType.APPLICATION_JSON_VALUE).
+				when().
+				post("/lines").
+				then().
+				log().all().
+				statusCode(HttpStatus.CREATED.value()).
+				extract().as(LineResponse.class);
+	}
 
-    public LineDetailResponse getLine(Long id) {
-        return
-            given().when().
-                get("/lines/" + id).
-                then().
-                log().all().
-                extract().as(LineDetailResponse.class);
-    }
+	public LineDetailResponse getLine(Long id) {
+		return
+			given().when().
+				get("/lines/" + id).
+				then().
+				log().all().
+				extract().as(LineDetailResponse.class);
+	}
 
-    public void updateLine(Long id, LocalTime startTime, LocalTime endTime) {
-        Map<String, String> params = new HashMap<>();
-        params.put("startTime", startTime.format(DateTimeFormatter.ISO_LOCAL_TIME));
-        params.put("endTime", endTime.format(DateTimeFormatter.ISO_LOCAL_TIME));
-        params.put("intervalTime", "10");
+	public void updateLine(Long id, LocalTime startTime, LocalTime endTime) {
+		Map<String, String> params = new HashMap<>();
+		params.put("startTime", startTime.format(DateTimeFormatter.ISO_LOCAL_TIME));
+		params.put("endTime", endTime.format(DateTimeFormatter.ISO_LOCAL_TIME));
+		params.put("intervalTime", "10");
 
-        given().
-            body(params).
-            contentType(MediaType.APPLICATION_JSON_VALUE).
-            accept(MediaType.APPLICATION_JSON_VALUE).
-            when().
-            put("/lines/" + id).
-            then().
-            log().all().
-            statusCode(HttpStatus.OK.value());
-    }
+		given().
+			body(params).
+			contentType(MediaType.APPLICATION_JSON_VALUE).
+			accept(MediaType.APPLICATION_JSON_VALUE).
+			when().
+			put("/lines/" + id).
+			then().
+			log().all().
+			statusCode(HttpStatus.OK.value());
+	}
 
-    public List<LineResponse> getLines() {
-        return
-            given().when().
-                get("/lines").
-                then().
-                log().all().
-                extract().
-                jsonPath().getList(".", LineResponse.class);
-    }
+	public List<LineResponse> getLines() {
+		return
+			given().when().
+				get("/lines").
+				then().
+				log().all().
+				extract().
+				jsonPath().getList(".", LineResponse.class);
+	}
 
-    public void deleteLine(Long id) {
-        given().when().
-            delete("/lines/" + id).
-            then().
-            log().all();
-    }
+	public void deleteLine(Long id) {
+		given().when().
+			delete("/lines/" + id).
+			then().
+			log().all();
+	}
 
-    public void addLineStation(Long lineId, Long preStationId, Long stationId) {
-        addLineStation(lineId, preStationId, stationId, 10, 10);
-    }
+	public void addLineStation(Long lineId, Long preStationId, Long stationId) {
+		addLineStation(lineId, preStationId, stationId, 10, 10);
+	}
 
-    public void addLineStation(Long lineId, Long preStationId, Long stationId, Integer distance,
-        Integer duration) {
-        Map<String, String> params = new HashMap<>();
-        params.put("preStationId", preStationId == null ? "" : preStationId.toString());
-        params.put("stationId", stationId.toString());
-        params.put("distance", distance.toString());
-        params.put("duration", duration.toString());
+	public void addLineStation(Long lineId, Long preStationId, Long stationId, Integer distance,
+		Integer duration) {
+		Map<String, String> params = new HashMap<>();
+		params.put("preStationId", preStationId == null ? "" : preStationId.toString());
+		params.put("stationId", stationId.toString());
+		params.put("distance", distance.toString());
+		params.put("duration", duration.toString());
 
-        given().
-            body(params).
-            contentType(MediaType.APPLICATION_JSON_VALUE).
-            accept(MediaType.APPLICATION_JSON_VALUE).
-            when().
-            post("/lines/" + lineId + "/stations").
-            then().
-            log().all().
-            statusCode(HttpStatus.OK.value());
-    }
+		given().
+			body(params).
+			contentType(MediaType.APPLICATION_JSON_VALUE).
+			accept(MediaType.APPLICATION_JSON_VALUE).
+			when().
+			post("/lines/" + lineId + "/stations").
+			then().
+			log().all().
+			statusCode(HttpStatus.OK.value());
+	}
 
-    public void removeLineStation(Long lineId, Long stationId) {
-        given().
-            contentType(MediaType.APPLICATION_JSON_VALUE).
-            accept(MediaType.APPLICATION_JSON_VALUE).
-            when().
-            delete("/lines/" + lineId + "/stations/" + stationId).
-            then().
-            log().all().
-            statusCode(HttpStatus.NO_CONTENT.value());
-    }
+	public void removeLineStation(Long lineId, Long stationId) {
+		given().
+			contentType(MediaType.APPLICATION_JSON_VALUE).
+			accept(MediaType.APPLICATION_JSON_VALUE).
+			when().
+			delete("/lines/" + lineId + "/stations/" + stationId).
+			then().
+			log().all().
+			statusCode(HttpStatus.NO_CONTENT.value());
+	}
 
-    public WholeSubwayResponse retrieveWholeSubway() {
-        return
-            given().
-                when().
-                get("/lines/detail").
-                then().
-                log().all().
-                extract().as(WholeSubwayResponse.class);
-    }
+	public WholeSubwayResponse retrieveWholeSubway() {
+		return
+			given().
+				when().
+				get("/lines/detail").
+				then().
+				log().all().
+				extract().as(WholeSubwayResponse.class);
+	}
 
-    public PathResponse findPath(String source, String target, String type) {
-        return
-            given().
-                contentType(MediaType.APPLICATION_JSON_VALUE).
-                accept(MediaType.APPLICATION_JSON_VALUE).
-                when().
-                get("/paths?source=" + source + "&target=" + target + "&type=" + type).
-                then().
-                log().all().
-                statusCode(HttpStatus.OK.value()).
-                extract().as(PathResponse.class);
-    }
+	public PathResponse findPath(String source, String target, String type) {
+		return
+			given().
+				contentType(MediaType.APPLICATION_JSON_VALUE).
+				accept(MediaType.APPLICATION_JSON_VALUE).
+				when().
+				get("/paths?source=" + source + "&target=" + target + "&type=" + type).
+				then().
+				log().all().
+				statusCode(HttpStatus.OK.value()).
+				extract().as(PathResponse.class);
+	}
 
-    /**
-     * 강남 - 역삼 - 선릉
-     * |           |
-     * |          한티
-     * |           |
-     * 양재 - 매봉 - 도곡
-     */
-    public void initStation() {
-        // 역 등록
-        StationResponse stationResponse1 = createStation(STATION_NAME_KANGNAM);
-        StationResponse stationResponse2 = createStation(STATION_NAME_YEOKSAM);
-        StationResponse stationResponse3 = createStation(STATION_NAME_SEOLLEUNG);
-        StationResponse stationResponse4 = createStation(STATION_NAME_HANTI);
-        StationResponse stationResponse5 = createStation(STATION_NAME_DOGOK);
-        StationResponse stationResponse6 = createStation(STATION_NAME_MAEBONG);
-        StationResponse stationResponse7 = createStation(STATION_NAME_YANGJAE);
+	/**
+	 * 강남 - 역삼 - 선릉
+	 * |           |
+	 * |          한티
+	 * |           |
+	 * 양재 - 매봉 - 도곡
+	 */
+	public void initStation() {
+		// 역 등록
+		StationResponse stationResponse1 = createStation(STATION_NAME_KANGNAM);
+		StationResponse stationResponse2 = createStation(STATION_NAME_YEOKSAM);
+		StationResponse stationResponse3 = createStation(STATION_NAME_SEOLLEUNG);
+		StationResponse stationResponse4 = createStation(STATION_NAME_HANTI);
+		StationResponse stationResponse5 = createStation(STATION_NAME_DOGOK);
+		StationResponse stationResponse6 = createStation(STATION_NAME_MAEBONG);
+		StationResponse stationResponse7 = createStation(STATION_NAME_YANGJAE);
 
-        // 2호선
-        LineResponse lineResponse1 = createLine("2호선");
-        addLineStation(lineResponse1.getId(), null, stationResponse1.getId(), 0, 0);
-        addLineStation(lineResponse1.getId(), stationResponse1.getId(), stationResponse2.getId(), 5,
-            10);
-        addLineStation(lineResponse1.getId(), stationResponse2.getId(), stationResponse3.getId(), 5,
-            10);
+		// 2호선
+		LineResponse lineResponse1 = createLine("2호선");
+		addLineStation(lineResponse1.getId(), null, stationResponse1.getId(), 0, 0);
+		addLineStation(lineResponse1.getId(), stationResponse1.getId(), stationResponse2.getId(), 5,
+			10);
+		addLineStation(lineResponse1.getId(), stationResponse2.getId(), stationResponse3.getId(), 5,
+			10);
 
-        // 분당선
-        LineResponse lineResponse2 = createLine("분당선");
-        addLineStation(lineResponse2.getId(), null, stationResponse3.getId(), 0, 0);
-        addLineStation(lineResponse2.getId(), stationResponse3.getId(), stationResponse4.getId(), 5,
-            10);
-        addLineStation(lineResponse2.getId(), stationResponse4.getId(), stationResponse5.getId(), 5,
-            10);
+		// 분당선
+		LineResponse lineResponse2 = createLine("분당선");
+		addLineStation(lineResponse2.getId(), null, stationResponse3.getId(), 0, 0);
+		addLineStation(lineResponse2.getId(), stationResponse3.getId(), stationResponse4.getId(), 5,
+			10);
+		addLineStation(lineResponse2.getId(), stationResponse4.getId(), stationResponse5.getId(), 5,
+			10);
 
-        // 3호선
-        LineResponse lineResponse3 = createLine("3호선");
-        addLineStation(lineResponse3.getId(), null, stationResponse5.getId(), 0, 0);
-        addLineStation(lineResponse3.getId(), stationResponse5.getId(), stationResponse6.getId(), 5,
-            10);
-        addLineStation(lineResponse3.getId(), stationResponse6.getId(), stationResponse7.getId(), 5,
-            10);
+		// 3호선
+		LineResponse lineResponse3 = createLine("3호선");
+		addLineStation(lineResponse3.getId(), null, stationResponse5.getId(), 0, 0);
+		addLineStation(lineResponse3.getId(), stationResponse5.getId(), stationResponse6.getId(), 5,
+			10);
+		addLineStation(lineResponse3.getId(), stationResponse6.getId(), stationResponse7.getId(), 5,
+			10);
 
-        // 신분당선
-        LineResponse lineResponse4 = createLine("신분당선");
-        addLineStation(lineResponse4.getId(), null, stationResponse1.getId(), 0, 0);
-        addLineStation(lineResponse4.getId(), stationResponse1.getId(), stationResponse7.getId(),
-            40, 3);
-    }
+		// 신분당선
+		LineResponse lineResponse4 = createLine("신분당선");
+		addLineStation(lineResponse4.getId(), null, stationResponse1.getId(), 0, 0);
+		addLineStation(lineResponse4.getId(), stationResponse1.getId(), stationResponse7.getId(),
+			40, 3);
+	}
 
-    public int createMember(String email, String name, String password) {
-        Map<String, String> params = new HashMap<>();
-        params.put("email", email);
-        params.put("name", name);
-        params.put("password", password);
+	public int createMember(String email, String name, String password) {
+		Map<String, String> params = new HashMap<>();
+		params.put("email", email);
+		params.put("name", name);
+		params.put("password", password);
 
-        return given().
-            body(params).
-            contentType(MediaType.APPLICATION_JSON_VALUE).
-            accept(MediaType.APPLICATION_JSON_VALUE).
-            when().
-            post("/me").
-            then().
-            log().all().
-            statusCode(HttpStatus.NO_CONTENT.value()).
-            extract().statusCode();
-    }
+		return given().
+			body(params).
+			contentType(MediaType.APPLICATION_JSON_VALUE).
+			accept(MediaType.APPLICATION_JSON_VALUE).
+			when().
+			post("/me").
+			then().
+			log().all().
+			statusCode(HttpStatus.NO_CONTENT.value()).
+			extract().statusCode();
+	}
 
-    public MemberResponse getMember(String email) {
-        return
-            given().
-                accept(MediaType.APPLICATION_JSON_VALUE).
-                when().
-                get("/members?email=" + email).
-                then().
-                log().all().
-                statusCode(HttpStatus.OK.value()).
-                extract().as(MemberResponse.class);
-    }
+	public MemberResponse getMember(String email) {
+		return
+			given().
+				accept(MediaType.APPLICATION_JSON_VALUE).
+				when().
+				get("/members?email=" + email).
+				then().
+				log().all().
+				statusCode(HttpStatus.OK.value()).
+				extract().as(MemberResponse.class);
+	}
 
-    public MemberResponse getMember(String email, TokenResponse tokenResponse
-    ) {
-        return
-            given().
-                header("Authorization",
-                    tokenResponse.getTokenType() + " " + tokenResponse.getAccessToken())
-                .accept(MediaType.APPLICATION_JSON_VALUE)
-                .when()
-                // .sessionId("loginMemberEmail", email)
-                .get("/me")
-                .then()
-                .log().all()
-                .statusCode(HttpStatus.OK.value())
-                .extract().as(MemberResponse.class);
-    }
+	public MemberResponse getMember(String email, TokenResponse tokenResponse
+	) {
+		return
+			given().
+				header("Authorization",
+					tokenResponse.getTokenType() + " " + tokenResponse.getAccessToken())
+				.accept(MediaType.APPLICATION_JSON_VALUE)
+				.when()
+				// .sessionId("loginMemberEmail", email)
+				.get("/me")
+				.then()
+				.log().all()
+				.statusCode(HttpStatus.OK.value())
+				.extract().as(MemberResponse.class);
+	}
 
-    public void updateMember(MemberResponse memberResponse) {
-        Map<String, String> params = new HashMap<>();
-        params.put("name", "NEW_" + TEST_USER_NAME);
-        params.put("password", "NEW_" + TEST_USER_PASSWORD);
+	public void updateMember(MemberResponse memberResponse) {
+		Map<String, String> params = new HashMap<>();
+		params.put("name", "NEW_" + TEST_USER_NAME);
+		params.put("password", "NEW_" + TEST_USER_PASSWORD);
 
-        given().
-            body(params).
-            contentType(MediaType.APPLICATION_JSON_VALUE).
-            accept(MediaType.APPLICATION_JSON_VALUE).
-            when().
-            put("/me").
-            then().
-            log().all().
-            statusCode(HttpStatus.OK.value());
-    }
+		given().
+			body(params).
+			contentType(MediaType.APPLICATION_JSON_VALUE).
+			accept(MediaType.APPLICATION_JSON_VALUE).
+			when().
+			put("/me").
+			then().
+			log().all().
+			statusCode(HttpStatus.OK.value());
+	}
 
-    public void deleteMember(TokenResponse tokenResponse) {
-        given().
-            header("Authorization",
-                tokenResponse.getTokenType() + " " + tokenResponse.getAccessToken())
-            .when()
-            .delete("/me")
-            .then()
-            .log().all()
-            .statusCode(HttpStatus.NO_CONTENT.value());
-    }
+	public void deleteMember(TokenResponse tokenResponse) {
+		given().
+			header("Authorization",
+				tokenResponse.getTokenType() + " " + tokenResponse.getAccessToken())
+			.when()
+			.delete("/me")
+			.then()
+			.log().all()
+			.statusCode(HttpStatus.NO_CONTENT.value());
+	}
 
-    public TokenResponse login(String email, String password) {
-        Map<String, String> paramMap = new HashMap<>();
-        paramMap.put("email", email);
-        paramMap.put("password", password);
+	public TokenResponse login(String email, String password) {
+		Map<String, String> paramMap = new HashMap<>();
+		paramMap.put("email", email);
+		paramMap.put("password", password);
 
-        return
-            given()
-                .body(paramMap)
-                .contentType(MediaType.APPLICATION_JSON_VALUE)
-                .accept(MediaType.APPLICATION_JSON_VALUE)
-                .when()
-                .post("/me/login")
-                .then()
-                .log().all()
-                .statusCode(HttpStatus.OK.value())
-                .extract().as(TokenResponse.class);
-    }
+		return
+			given()
+				.body(paramMap)
+				.contentType(MediaType.APPLICATION_JSON_VALUE)
+				.accept(MediaType.APPLICATION_JSON_VALUE)
+				.when()
+				.post("/me/login")
+				.then()
+				.log().all()
+				.statusCode(HttpStatus.OK.value())
+				.extract().as(TokenResponse.class);
+	}
 
-    public void updateMemberWhenLoggedIn(String accessToken,
-        String name,
-        String password) {
-        Map<String, String> params = new HashMap<>();
-        params.put("name", name);
-        params.put("password", password);
+	public void updateMemberWhenLoggedIn(String accessToken,
+		String name,
+		String password) {
+		Map<String, String> params = new HashMap<>();
+		params.put("name", name);
+		params.put("password", password);
 
-        given().header("Authorization", "bearer " + accessToken);
+		given().header("Authorization", "bearer " + accessToken);
 
-        given()
-            .header("Authorization", "bearer " + accessToken)
-            .body(params)
-            .contentType(MediaType.APPLICATION_JSON_VALUE)
-            .accept(MediaType.APPLICATION_JSON_VALUE)
-            .when()
-            .put("/me")
-            .then()
-            .log().all()
-            .statusCode(HttpStatus.NO_CONTENT.value());
-    }
+		given()
+			.header("Authorization", "bearer " + accessToken)
+			.body(params)
+			.contentType(MediaType.APPLICATION_JSON_VALUE)
+			.accept(MediaType.APPLICATION_JSON_VALUE)
+			.when()
+			.put("/me")
+			.then()
+			.log().all()
+			.statusCode(HttpStatus.NO_CONTENT.value());
+	}
 }
